EDSC-4055: Replacing tests back
